fix cardbacks, I think? + add working migration script
